+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.5] - 09-23-2021
+
+* Updating the hex color validator. The validation mode can now be specified (with #, without or both).
+* Adding the possibility to disable the final action with the AFWidget.
+
 ## [0.2.4] - 09-23-2021
 
 * Exporting the validator.dart file inside the library main file.